Make generic (modular) tests for datamodules
